requirements.txt
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.127.0
anthropic==0.49.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
